govtech/scripts/requirements.txt
Django
mysqlclient
validate-email-address
pillow
pymysql
django-debug-toolbar
djangorestframework
djangorestframework-simplejwt
sendgrid
psycopg2-binary
